+++ contrib/msggen/msggen/utils/__init__.py
@@ -1,0 +1,1 @@
+from .utils import load_jsonrpc_method, load_jsonrpc_service, repo_root # noqa
